+++ CHANGELOG.md
@@ -4,7 +4,8 @@ Changelog
 **Next release**
 
 * Mark all functions in the system category as dangerous, except one
-* Don't return an empty gids list
+* Don't return an empty gids list to prevent exception
+* Require [Pro enabler](http://play.google.com/store/apps/details?id=biz.bokhorst.xprivacy.pro) version 1.12
 
 **Version 1.10.20 TEST**